Conflicts:
	eds/api/src/main/java/com/edaisong/api/dao/impl/OrderGrabChildDao.java
	eds/api/src/main/java/com/edaisong/api/service/impl/OrderGrabService.java

+++ eds/api/src/main/java/com/edaisong/api/service/inter/IOrderSubsidiesLogService.java
@@ -12,5 +12,5 @@ public interface IOrderSubsidiesLogService {
      * @date 20150827  
      * @return
      */
-   List<OrderSubsidiesLog>  GetOrderOptionLog(int orderId);
+   List<OrderSubsidiesLog>  GetOrderOptionLog(Long orderId);
 }
